+++ packages/xcm-cfg/src/configs/kusama/index.ts
@@ -4,7 +4,6 @@ import { assetHubConfig } from './assethub';
 import { basiliskConfig } from './basilisk';
 import { integriteeConfig } from './integritee';
 import { karuraConfig } from './karura';
-import { kusamaConfig } from './kusama';
 import { tinkernetConfig } from './tinkernet';
 import { robonomicsConfig } from './robonomics';
 
@@ -13,7 +12,6 @@ export const kusamaChainsConfig: ChainRoutes[] = [
   basiliskConfig,
   integriteeConfig,
   karuraConfig,
-  kusamaConfig,
   tinkernetConfig,
   robonomicsConfig,
 ];