+++ packages/xcm-cfg/src/configs/polkadot/hydration/index.ts
@@ -85,21 +85,19 @@ import {
   neuroweb,
   nodle,
   pendulum,
-  polkadot,
   phala,
   solana,
   sui_chain,
   unique,
   zeitgeist,
-  polkadotCex,
-  xode,
+  xode
 } from '../../../chains';
 import { ExtrinsicBuilder, XcmTransferType } from '../../../builders';
 
 import { balance, fee } from './configs';
 import {
   toHubExtTemplate,
-  toHubWithCexFwd2Template,
+  toHubWithCexFwdTemplate,
   toMoonbeamErc20Template,
   toZeitgeistErc20Template,
   toTransferTemplate,
@@ -214,27 +212,6 @@ const toAstar: AssetRoute[] = [
   toTransferTemplate(vastr, astar, 0.005),
   toTransferTemplate(usdt, astar, 0.3),
   toTransferTemplate(usdc, astar, 0.3),
-  // new AssetRoute({
-  //   source: {
-  //     asset: dot,
-  //     balance: balance(),
-  //     fee: fee(),
-  //     destinationFee: {
-  //       balance: balance(),
-  //     },
-  //   },
-  //   destination: {
-  //     chain: astar,
-  //     asset: dot,
-  //     fee: {
-  //       amount: 0.1,
-  //       asset: dot,
-  //     },
-  //   },
-  //   extrinsic: ExtrinsicBuilder().polkadotXcm().transferAssetsUsingTypeAndThen({
-  //     transferType: XcmTransferType.RemoteReserve,
-  //   }),
-  // }),
 ];
 
 const toBifrost: AssetRoute[] = [
@@ -300,6 +277,27 @@ const toInterlay: AssetRoute[] = [
     },
     extrinsic: ExtrinsicBuilder().xTokens().transfer(),
   }),
+  new AssetRoute({
+    source: {
+      asset: dot,
+      balance: balance(),
+      fee: fee(),
+      destinationFee: {
+        balance: balance(),
+      },
+    },
+    destination: {
+      chain: interlay,
+      asset: dot,
+      fee: {
+        amount: 0.05,
+        asset: dot,
+      },
+    },
+    extrinsic: ExtrinsicBuilder().polkadotXcm().transferAssetsUsingTypeAndThen({
+      transferType: XcmTransferType.RemoteReserve,
+    }),
+  }),
 ];
 
 const toMoonbeam: AssetRoute[] = [
@@ -315,8 +313,6 @@ const toMoonbeam: AssetRoute[] = [
   toMoonbeamErc20Template(weth_mwh),
   toMoonbeamErc20Template(sol),
 ];
-
-const toPolkadot: AssetRoute[] = [toTransferTemplate(dot, polkadot, 0.003)];
 
 const toZeitgeist: AssetRoute[] = [
   toTransferTemplate(ztg, zeitgeist, 0.0093),
@@ -384,46 +380,24 @@ const toSuiViaWormhole: AssetRoute[] = [
   viaWormholeRelayerTemplate(sui, sui, sui_chain),
 ];
 
-const toCexViaRelay = new AssetRoute({
-  source: {
-    asset: dot,
-    balance: balance(),
-    fee: fee(),
-    destinationFee: {
-      balance: balance(),
-    },
-  },
-  destination: {
-    chain: polkadotCex,
-    asset: dot,
-    fee: {
-      amount: 1.1,
-      asset: dot,
-    },
-  },
-  extrinsic: ExtrinsicBuilder()
-    .utility()
-    .batchAll([
-      ExtrinsicBuilder().xTokens().transfer(),
-      ExtrinsicBuilder().polkadotXcm().send().transact({
-        fee: 0.01,
-      }),
-    ]),
-  transact: {
-    chain: polkadot,
-    fee: {
-      amount: 0,
-      asset: dot,
-      balance: balance(),
-    },
-    extrinsic: ExtrinsicBuilder().balances().transferAll(),
-  },
-});
-
 const toCex: AssetRoute[] = [
-  toHubWithCexFwd2Template(usdt),
-  toHubWithCexFwd2Template(usdc),
-  toCexViaRelay,
+  toHubWithCexFwdTemplate(
+    usdt,
+    0.1,
+    ExtrinsicBuilder().xTokens().transferMultiasset()
+  ),
+  toHubWithCexFwdTemplate(
+    usdc,
+    0.1,
+    ExtrinsicBuilder().xTokens().transferMultiasset()
+  ),
+  toHubWithCexFwdTemplate(
+    dot,
+    0.2,
+    ExtrinsicBuilder().polkadotXcm().transferAssetsUsingTypeAndThen({
+      transferType: XcmTransferType.DestinationReserve,
+    })
+  ),
 ];
 
 const toXode: AssetRoute[] = [
@@ -453,7 +427,6 @@ export const hydrationConfig = new ChainRoutes({
     ...toNeuroweb,
     ...toNodle,
     ...toPhala,
-    ...toPolkadot,
     ...toPendulum,
     ...toSolanaViaWormhole,
     ...toSuiViaWormhole,