+++ packages/xcm-cfg/src/configs/polkadot/interlay.ts
@@ -133,6 +133,28 @@ const toHydration: AssetRoute[] = [
     },
     extrinsic: ExtrinsicBuilder().xTokens().transfer(),
   }),
+  new AssetRoute({
+    source: {
+      asset: dot,
+      balance: BalanceBuilder().substrate().tokens().accounts(),
+      fee: {
+        asset: intr,
+        balance: BalanceBuilder().substrate().tokens().accounts(),
+      },
+      destinationFee: {
+        balance: BalanceBuilder().substrate().tokens().accounts(),
+      },
+    },
+    destination: {
+      chain: hydration,
+      asset: dot,
+      fee: {
+        amount: 0.1,
+        asset: dot,
+      },
+    },
+    extrinsic: ExtrinsicBuilder().xTokens().transfer(),
+  }),
 ];
 
 export const interlayConfig = new ChainRoutes({